style: 바인딩 어댑터 분리
